10344: Split Grant/Deny motion action tests into movant and shared fixture files
